gst-plugins-renderer: autotools compiling
This patch changes the way of the compiling the library using autotools.
The previous 'Makefile' is renamed to 'Makefile.cross'.
and the libary needs a extra dependency path for gstreamer.
so, the path('gst_path') is added using with option when we set a configuration.
(--with-gst_path=the dir where the libraries exist about gstreamer)

Change-Id: I7880fb81a7fe47400ade1fc1dbccc8c5af10fa70
Signed-off-by: HyejungKwon <cjscld15@nexell.co.kr>